requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.5.1
torchvision==0.20.1
torchaudio
transformers
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-learn
matplotlib
seaborn
tqdm
albumentations
ipykernel